[WK4] Implement read.

+++ internal/common/block.go
@@ -2,26 +2,26 @@
 // different packages.
 package common
 
-// Block represents a census block group in California.
+// Block represents a census block group.
 type Block struct {
 	// Value represents the median house value in the block.
-	Value float32
+	Value float64
 
 	// Income represents the median income of households in the block, in units of $10,000.
-	Income float32
+	Income float64
 
 	// Age indicates the median age of houses in the block.
-	Age float32
+	Age float64
 
 	// Rooms is the total number of rooms in the block.
-	Rooms float32
+	Rooms float64
 
 	// Bedrooms is the total number of bedrooms in the block.
-	Bedrooms float32
+	Bedrooms float64
 
 	// Population is the total number of people residing in the block.
-	Population float32
+	Population float64
 
 	// Households represents the total number of households in the block (?).
-	Households float32
+	Households float64
 }
